+++ client/src/main/java/fiit/vava/client/routes/admin/Controller.java
@@ -1,0 +1,4 @@
+package fiit.vava.client.routes.admin;
+
+public class Controller {
+}