+++ server/src/main/java/module-info.java
@@ -9,6 +9,8 @@ module server {
     requires com.google.common;
     requires jjwt;
 
+    requires org.slf4j;
+
     exports fiit.vava.server;
     exports fiit.vava.server.config;
 }